Rename note widget to note widget light
