+++ Simplenote/src/main/java/com/automattic/simplenote/NoteEditorActivity.java
@@ -27,7 +27,7 @@ import org.wordpress.passcodelock.AppLockManager;
 
 import java.util.ArrayList;
 
-import static com.automattic.simplenote.NoteWidget.KEY_WIDGET_CLICK;
+import static com.automattic.simplenote.NoteWidgetLight.KEY_WIDGET_CLICK;
 import static com.automattic.simplenote.analytics.AnalyticsTracker.CATEGORY_WIDGET;
 import static com.automattic.simplenote.analytics.AnalyticsTracker.Stat.NOTE_WIDGET_NOTE_TAPPED;
 import static com.automattic.simplenote.utils.DisplayUtils.disableScreenshotsIfLocked;
