+++ Simplenote/src/main/java/com/automattic/simplenote/NoteWidgetLight.java
@@ -25,14 +25,14 @@ import static com.automattic.simplenote.analytics.AnalyticsTracker.Stat.NOTE_WID
 import static com.automattic.simplenote.analytics.AnalyticsTracker.Stat.NOTE_WIDGET_NOTE_TAPPED;
 import static com.automattic.simplenote.analytics.AnalyticsTracker.Stat.NOTE_WIDGET_SIGN_IN_TAPPED;
 
-public class NoteWidget extends AppWidgetProvider {
+public class NoteWidgetLight extends AppWidgetProvider {
     public static final String KEY_WIDGET_CLICK = "key_widget_click";
-    public static final String KEY_WIDGET_IDS = "key_widget_ids";
+    public static final String KEY_WIDGET_IDS_LIGHT = "key_widget_ids_light";
 
     @Override
     public void onReceive(Context context, Intent intent) {
-        if (intent.getExtras() != null && intent.hasExtra(KEY_WIDGET_IDS)) {
-            int[] ids = intent.getExtras().getIntArray(KEY_WIDGET_IDS);
+        if (intent.getExtras() != null && intent.hasExtra(KEY_WIDGET_IDS_LIGHT)) {
+            int[] ids = intent.getExtras().getIntArray(KEY_WIDGET_IDS_LIGHT);
             this.onUpdate(context, AppWidgetManager.getInstance(context), ids);
         } else {
             super.onReceive(context, intent);
@@ -49,7 +49,7 @@ public class NoteWidget extends AppWidgetProvider {
 
     @Override
     public void onAppWidgetOptionsChanged(Context context, AppWidgetManager appWidgetManager, int appWidgetId, Bundle newOptions) {
-        RemoteViews views = new RemoteViews(context.getPackageName(), R.layout.note_widget);
+        RemoteViews views = new RemoteViews(context.getPackageName(), R.layout.note_widget_light);
         resizeWidget(newOptions, views);
         appWidgetManager.updateAppWidget(appWidgetId, views);
     }
@@ -96,7 +96,7 @@ public class NoteWidget extends AppWidgetProvider {
 
     private void updateWidget(Context context, AppWidgetManager appWidgetManager, int appWidgetId, Bundle appWidgetOptions) {
         // Get widget views
-        RemoteViews views = new RemoteViews(context.getPackageName(), R.layout.note_widget);
+        RemoteViews views = new RemoteViews(context.getPackageName(), R.layout.note_widget_light);
         resizeWidget(appWidgetOptions, views);
 
         // Verify user authentication
@@ -105,7 +105,7 @@ public class NoteWidget extends AppWidgetProvider {
         User user = simperium.getUser();
 
         if (user.getStatus().equals(User.Status.NOT_AUTHORIZED)) {
-            // Create intent to navigate to notes activity which redirects to signin on widget click
+            // Create intent to navigate to notes activity which redirects to login on widget click
             Intent intent = new Intent(context, NotesActivity.class);
             intent.putExtra(KEY_WIDGET_CLICK, NOTE_WIDGET_SIGN_IN_TAPPED);
             intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TOP);
@@ -157,7 +157,7 @@ public class NoteWidget extends AppWidgetProvider {
                     views.setTextViewText(R.id.widget_text_content, content);
                 } catch (BucketObjectMissingException e) {
                     // Create intent to navigate to widget configure activity on widget click
-                    Intent intent = new Intent(context, NoteWidgetConfigureActivity.class);
+                    Intent intent = new Intent(context, NoteWidgetLightConfigureActivity.class);
                     intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TOP);
                     intent.putExtra(KEY_WIDGET_CLICK, NOTE_WIDGET_NOTE_NOT_FOUND_TAPPED);
                     intent.putExtra(AppWidgetManager.EXTRA_APPWIDGET_ID, appWidgetId);
