+++ Simplenote/src/main/java/com/automattic/simplenote/utils/WidgetUtils.java
@@ -5,15 +5,16 @@ import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
 
-import com.automattic.simplenote.NoteWidget;
+import com.automattic.simplenote.NoteWidgetLight;
 
 public class WidgetUtils {
     public static void updateNoteWidgets(Context context) {
         AppWidgetManager appWidgetManager = AppWidgetManager.getInstance(context);
-        int[] ids = appWidgetManager.getAppWidgetIds(new ComponentName(context, NoteWidget.class));
-        Intent updateIntent = new Intent();
-        updateIntent.setAction(AppWidgetManager.ACTION_APPWIDGET_UPDATE);
-        updateIntent.putExtra(NoteWidget.KEY_WIDGET_IDS, ids);
-        context.sendBroadcast(updateIntent);
+
+        int[] idsLight = appWidgetManager.getAppWidgetIds(new ComponentName(context, NoteWidgetLight.class));
+        Intent updateIntentLight = new Intent();
+        updateIntentLight.setAction(AppWidgetManager.ACTION_APPWIDGET_UPDATE);
+        updateIntentLight.putExtra(NoteWidgetLight.KEY_WIDGET_IDS_LIGHT, idsLight);
+        context.sendBroadcast(updateIntentLight);
     }
 }
