+++ Simplenote/src/main/res/drawable/note_widget_background_light.xml
@@ -4,7 +4,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android">
 
     <solid
-        android:color="@android:color/white">
+        android:color="@color/background_light">
     </solid>
 
     <corners
