+++ Simplenote/src/main/res/layout/note_widget_light.xml
@@ -4,7 +4,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/widget_layout"
-    android:background="@drawable/note_widget_background"
+    android:background="@drawable/note_widget_background_light"
     android:foreground="?android:attr/selectableItemBackgroundBorderless"
     android:layout_height="match_parent"
     android:layout_width="match_parent"
