+++ Simplenote/src/main/res/values/strings.xml
@@ -146,7 +146,7 @@
     <string name="font_size_extra_large">Extra large</string>
 
     <!-- Pinned Note Widget -->
-    <string name="label_note">Note</string>
+    <string name="note_widget_light">Note (Light)</string>
     <string name="loading_note">Loading note&#8230;</string>
     <string name="note_not_found">Note not found</string>
     <string name="select_note">Select Note</string>
