+++ Simplenote/src/main/res/xml/note_widget_info_light.xml
@@ -3,11 +3,11 @@
 <appwidget-provider
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    android:configure="com.automattic.simplenote.NoteWidgetConfigureActivity"
-    android:initialLayout="@layout/note_widget"
+    android:configure="com.automattic.simplenote.NoteWidgetLightConfigureActivity"
+    android:initialLayout="@layout/note_widget_light"
     android:minHeight="@dimen/cell_widget"
     android:minWidth="@dimen/cell_widget"
-    android:previewImage="@drawable/note_widget_preview"
+    android:previewImage="@drawable/note_widget_preview_light"
     android:resizeMode="horizontal|vertical"
     android:updatePeriodMillis="21600000"
     android:widgetCategory="home_screen"
